requirements.txt
torch==2.1.1
pillow==10.1.0
opencv_python==4.8.1.78